backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
sqlite3
python-pptx==0.6.23
reportlab==4.0.7
PyPDF2==3.0.1
google-generativeai
python-docx